+++ src/main/java/at/hannibal2/skyhanni/utils/SoundUtils.kt
@@ -1,0 +1,34 @@
+package at.hannibal2.skyhanni.utils
+
+import net.minecraft.client.Minecraft
+import net.minecraft.client.audio.ISound
+import net.minecraft.client.audio.PositionedSound
+import net.minecraft.client.audio.SoundCategory
+import net.minecraft.util.ResourceLocation
+
+object SoundUtils {
+    fun ISound.playSound() {
+        val gameSettings = Minecraft.getMinecraft().gameSettings
+        val oldLevel = gameSettings.getSoundLevel(SoundCategory.PLAYERS)
+        gameSettings.setSoundLevel(SoundCategory.PLAYERS, 1f)
+        try {
+            Minecraft.getMinecraft().soundHandler.playSound(this)
+        } catch (e: Exception) {
+            e.printStackTrace()
+        }
+        gameSettings.setSoundLevel(SoundCategory.PLAYERS, oldLevel)
+    }
+
+    fun createSound(name: String, pitch: Float): ISound {
+        val sound: ISound = object : PositionedSound(ResourceLocation(name)) {
+            init {
+                volume = 50f
+                repeat = false
+                repeatDelay = 0
+                attenuationType = ISound.AttenuationType.NONE
+                this.pitch = pitch
+            }
+        }
+        return sound
+    }
+}