+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Slayer
 + Added a cooldown when the Fire Pillars from the blaze slayer will kill you.
 + Added a faster and permanent display for the blaze slayer daggers
++ Added custom countdown sound for the Fire Pillar timer for the blaze slayer.
++ Added hide sound and entities when building the Fire Pillar for the blaze slayer.
 
 ### Misc
 + Added option to hide armor or just helmet of other player or yourself